refactor(cpp): huge refactor, move all exception classes to src, remove useless includes
