+++ cpp/include/sdk/Exception.h
@@ -19,7 +19,6 @@
 
 #include "MiraiCPMacros.h"
 // -----------------------
-#include "CPPPlugin.h"
 #include <exception>
 #include <string>
 #include <thread>
@@ -82,7 +81,7 @@ namespace MiraiCP {
     /// 继承该类后异常类能正确实现多态。
     /// @interface MiraiCPExceptionCRTP
     /// @note 请勿给该类增加新的属性。如果要增加属性应在 MiraiCPExceptionBase 中增加
-    template<class T>
+    template<typename T>
     class MIRAICP_EXPORT MiraiCPExceptionCRTP : public MiraiCPExceptionBase {
     public:
         /// 委托构造函数
@@ -98,215 +97,6 @@ namespace MiraiCP {
     /// @param const string &description, string _filename, int _lineNum
     /// @see MiraiCPExceptionBase
     class MIRAICP_EXPORT MiraiCPException : public MiraiCPExceptionCRTP<MiraiCPExceptionBase> {};
-
-    /// 文件读取异常.
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT UploadException : public MiraiCPExceptionCRTP<UploadException> {
-    public:
-        explicit UploadException(const std::string &text, string _filename, int _lineNum) : MiraiCPExceptionCRTP("上传(图片/文件)异常" + text, std::move(_filename), _lineNum) {}
-
-        static std::string exceptionType() { return "UploadException"; }
-    };
-
-    /// 通常为Mirai返回
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT IllegalStateException : public MiraiCPExceptionCRTP<IllegalStateException> {
-    public:
-        explicit IllegalStateException(const std::string &text, string _filename, int _lineNum) : MiraiCPExceptionCRTP("状态异常:" + text, std::move(_filename), _lineNum) {}
-
-        static std::string exceptionType() { return "IllegalStateException"; }
-    };
-
-    /// 内部异常, 通常为json读写问题
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT APIException : public MiraiCPExceptionCRTP<APIException> {
-    public:
-        explicit APIException(const std::string &text, string _filename, int _lineNum) : MiraiCPExceptionCRTP("MiraiCP内部无法预料的错误:" + text, std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "APIException"; }
-    };
-
-    /// 机器人操作异常
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT BotException : public MiraiCPExceptionCRTP<BotException> {
-    public:
-        explicit BotException(string _filename, int _lineNum) : MiraiCPExceptionCRTP("没有权限执行该操作", std::move(_filename), _lineNum) {}
-
-        explicit BotException(const string &d, string _filename, int _lineNum) : MiraiCPExceptionCRTP(d, std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "BotException"; }
-    };
-
-    /// 被禁言异常, 通常发生于发送信息
-    class MIRAICP_EXPORT BotIsBeingMutedException : public MiraiCPExceptionCRTP<BotIsBeingMutedException> {
-    public:
-        /// 剩余禁言时间, 单位秒
-        int timeRemain;
-
-    public:
-        explicit BotIsBeingMutedException(int t, string _filename, int _lineNum) : MiraiCPExceptionCRTP(
-                                                                                           "发送信息失败, bot已被禁言, 剩余时间" + std::to_string(t), std::move(_filename), _lineNum),
-                                                                                   timeRemain(t) {}
-
-        static string exceptionType() { return "BotIsBeingMutedException"; }
-    };
-
-    /// 禁言异常
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT MuteException : public MiraiCPExceptionCRTP<MuteException> {
-    public:
-        /*
-        *	 禁言时间超出0s~30d
-        */
-        MuteException(string _filename, int _lineNum) : MiraiCPExceptionCRTP("禁言时长不在0s~30d中间", std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "MuteException"; }
-    };
-
-    /// 获取群成员错误
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT MemberException : public MiraiCPExceptionCRTP<MemberException> {
-    public:
-        enum MemberExceptionType : int {
-            OtherType,
-            NoSuchGroup,
-            NoSuchMember
-        };
-        MemberExceptionType type = OtherType;
-
-        /*
-        *   "1" - 找不到群
-        *	"2" - 找不到群成员
-        */
-        explicit MemberException(int _type, string _filename, int _lineNum) : MiraiCPExceptionCRTP(
-                                                                                      [&]() -> string {
-                                                                                          type = MemberExceptionType(_type);
-                                                                                          switch (type) {
-                                                                                              case NoSuchGroup:
-                                                                                                  return "找不到群";
-                                                                                              case NoSuchMember:
-                                                                                                  return "找不到群成员";
-                                                                                              default:
-                                                                                                  return "";
-                                                                                          }
-                                                                                      }(),
-                                                                                      std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "MemberException"; }
-    };
-
-    /// 获取群成员错误
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT FriendException : public MiraiCPExceptionCRTP<FriendException> {
-    public:
-        /*
-        *   找不到好友
-        */
-        FriendException(string _filename, int _lineNum) : MiraiCPExceptionCRTP("找不到好友", std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "FriendException"; }
-    };
-
-    /// 获取群错误
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT GroupException : public MiraiCPExceptionCRTP<GroupException> {
-    public:
-        GroupException(string _filename, int _lineNum) : MiraiCPExceptionCRTP("找不到群", std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "GroupException"; }
-    };
-
-    /// 撤回异常
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT RecallException : public MiraiCPExceptionCRTP<RecallException> {
-    public:
-        RecallException(string _filename, int _lineNum) : MiraiCPExceptionCRTP("该消息已经被撤回", std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "RecallException"; }
-    };
-
-    /// 远程资源出现问题
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT RemoteAssetException : public MiraiCPExceptionCRTP<RemoteAssetException> {
-    public:
-        explicit RemoteAssetException(const string &e, string _filename, int _lineNum) : MiraiCPExceptionCRTP(e, std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "RemoteAssetException"; }
-    };
-
-    /// 参数错误
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT IllegalArgumentException : public MiraiCPExceptionCRTP<IllegalArgumentException> {
-    public:
-        explicit IllegalArgumentException(const string &e, string _filename, int _lineNum) : MiraiCPExceptionCRTP(e, std::move(_filename), _lineNum) {
-        }
-
-        static string exceptionType() { return "IllegalArgumentException"; }
-    };
-
-    /// 超时
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT TimeOutException : public MiraiCPExceptionCRTP<TimeOutException> {
-    public:
-        explicit TimeOutException(const std::string &e, string _filename, int _lineNum) : MiraiCPExceptionCRTP(e, std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "TimeOutException"; }
-    };
-
-    /// 事件被取消, 一般出现在发送消息时在preSendMessageEvent取消的时候抛出
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT EventCancelledException : public MiraiCPExceptionCRTP<EventCancelledException> {
-    public:
-        explicit EventCancelledException(const string &msg, string _filename, int _lineNum) : MiraiCPExceptionCRTP(msg, std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "EventCancelledException"; }
-    };
-
-    /// 插件没有权限时抛出该异常
-    /// 该异常仅可能在插件尝试调用libLoader 高级权限的Api接口时抛出
-    /// 如插件尝试重载、加载、卸载插件等操作，但配置文件中并没有赋予该插件权限时
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT PluginNotAuthorizedException : public MiraiCPExceptionCRTP<PluginNotAuthorizedException> {
-    public:
-        explicit PluginNotAuthorizedException(string _filename, int _lineNum) : MiraiCPExceptionCRTP("插件" + CPPPlugin::config.getId() + "没有管理权限", std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "PluginNotAuthorizedException"; }
-    };
-
-    /// 插件未加载抛出该异常
-    /// 在插件能正常运行时不会抛出，出现该异常事件时请不要再次尝试收发消息等Mirai操作，
-    /// 否则可能导致异常处理时再次抛出异常
-    /// @see MiraiCPExceptionBase
-    class MIRAICP_EXPORT PluginNotEnabledException : public MiraiCPExceptionCRTP<PluginNotEnabledException> {
-    public:
-        explicit PluginNotEnabledException(string _filename, int _lineNum) : MiraiCPExceptionCRTP("插件" + CPPPlugin::config.getId() + "未加载", std::move(_filename), _lineNum) {}
-
-        static string exceptionType() { return "PluginNotEnabledException"; }
-    };
-
-    /// 如果在 MiraiCPNewThread 中捕获到了非 MiraiCP 之外的异常抛出
-    /// @see MiraiCPNewThread
-    class MiraiCPThreadException : public MiraiCPExceptionCRTP<MiraiCPThreadException> {
-    public:
-        /// 抛出异常的线程 ID
-        std::thread::id threadId;
-
-    public:
-        explicit MiraiCPThreadException(const std::string &exception_content, std::thread::id threadId, string _filename, int _lineNum)
-            : MiraiCPExceptionCRTP(exception_content + " at threadId: " + getThreadIdStr(threadId), std::move(_filename), _lineNum),
-              threadId(threadId) {}
-
-    public:
-        std::string getThreadIdStr() const { return getThreadIdStr(threadId); }
-
-    public:
-        static string exceptionType() { return "MiraiCPThreadException"; }
-
-    private:
-        MIRAICP_EXPORT static std::string getThreadIdStr(const std::thread::id &id);
-    };
-
-    void ErrorHandle0(const std::string &name, int line, const std::string &re, const std::string &ErrorMsg = "");
 } // namespace MiraiCP
 
 #endif //MIRAICP_PRO_EXCEPTION_H
