+++ cpp/include/sdk/MessageChain.h
@@ -19,9 +19,8 @@
 
 #include "MiraiCPMacros.h"
 // -----------------------
-#include "Exception.h"
 #include "SingleMessage.h"
-#include "commonTools.h"
+#include <functional>
 
 
 namespace MiraiCP {
@@ -30,12 +29,12 @@ namespace MiraiCP {
     namespace internal {
         /// @brief 为 std::shared_ptr<SingleMessage> 增加功能的封装类
         /// @note dev: 不可加入其他成员变量
-        class Message : public std::shared_ptr<SingleMessage> {
+        class MIRAICP_EXPORT Message : public std::shared_ptr<SingleMessage> {
             typedef std::shared_ptr<SingleMessage> Super;
 
         public:
 #if MIRAICP_MSVC
-            Message() {} // for MSVC compatible, Message should be default constructable. See MessageChain
+            Message() {} /// for MSVC compatible, Message should be default constructable. See MessageChain
 #endif
 
             template<typename T, typename = std::enable_if<std::is_base_of_v<SingleMessage, T>>>
@@ -53,22 +52,22 @@ namespace MiraiCP {
 
             /// 取指定类型
             /// @throw IllegalArgumentException
-            template<class T>
+            /// @note 将复制一次对象
+            template<typename T>
             T getVal() const {
-                // for dev: 不用 get 为了不和shared_ptr重叠
+                return getRef<T>();
+            }
+
+            /// 取指定类型的const引用
+            /// @throw IllegalArgumentException
+            template<typename T>
+            const T &getRef() const {
+                assert(get());
                 static_assert(std::is_base_of_v<SingleMessage, T>, "只支持SingleMessage的派生类");
-                if (T::type() != getType())
-                    throw IllegalArgumentException(
-                            "cannot convert from " + SingleMessage::getTypeString(getType()) + " to " +
-                                    SingleMessage::getTypeString(T::type()),
-                            MIRAICP_EXCEPTION_WHERE);
-                T *re = static_cast<T *>(std::shared_ptr<SingleMessage>::get());
-                if (re == nullptr)
-                    throw IllegalArgumentException(
-                            "cannot convert from " + SingleMessage::getTypeString(getType()) + " to " +
-                                    SingleMessage::getTypeString(T::type()),
-                            MIRAICP_EXCEPTION_WHERE);
-                return *re;
+                if (T::type() != getType()) {
+                    messageThrow(SingleMessage::getTypeString(getType()), SingleMessage::getTypeString(T::type()), MIRAICP_EXCEPTION_WHERE);
+                }
+                return *static_cast<T *>(Super::get()); // safe cast!
             }
 
             [[nodiscard]] std::string toMiraiCode() const;
@@ -77,6 +76,9 @@ namespace MiraiCP {
 
             bool operator==(const Message &m) const;
             bool operator!=(const Message &m) const;
+
+        private:
+            static void messageThrow(const std::string &from, const std::string &to, const char *file, int line);
         };
     } // namespace internal
 
@@ -117,15 +119,12 @@ namespace MiraiCP {
         };
 
     public:
-        [[deprecated("MessageChain继承自std::vector<Message>，无需获取内部vector")]] const std::vector<Message> &vector() const {
-            return static_cast<const std::vector<Message> &>(*this);
-        }
-
         [[nodiscard]] std::string toMiraiCode() const override;
 
         [[nodiscard]] std::vector<std::string> toMiraiCodeVector() const;
 
         [[nodiscard]] nlohmann::json toJson() const;
+
         /**
          * @ensure toJson().dump()
          * @return MessageChain serialize to String
@@ -147,7 +146,8 @@ namespace MiraiCP {
         }
 
         /// 筛选出某种类型的消息
-        template<class T>
+        /// @note 最多可能将整个vector复制一次
+        template<typename T>
         std::vector<T> filter() {
             static_assert(std::is_base_of_v<SingleMessage, T>, "只支持SingleMessage的子类");
             std::vector<T> re;
@@ -159,7 +159,8 @@ namespace MiraiCP {
         }
 
         /// 自定义筛选器
-        template<class T>
+        /// @note 最多可能将整个vector复制一次
+        template<typename T>
         std::vector<T> filter(const std::function<bool(const Message &)> &func) {
             static_assert(std::is_base_of_v<SingleMessage, T>, "只支持SingleMessage的子类");
             std::vector<T> re;
@@ -171,7 +172,7 @@ namespace MiraiCP {
         }
 
         /// 找出第一个指定的type的消息，消息可能不存在
-        template<class T>
+        template<typename T>
         std::optional<T> first() {
             for (auto &&a: *this)
                 if (a.getType() == T::type())
@@ -179,7 +180,7 @@ namespace MiraiCP {
             return std::nullopt;
         }
 
-        template<class T>
+        template<typename T>
         [[nodiscard]] MessageChain plus(const T &a) const {
             static_assert(std::is_base_of_v<SingleMessage, T>, "只支持SingleMessage的子类");
             MessageChain tmp(*this);
@@ -201,19 +202,6 @@ namespace MiraiCP {
         bool operator!=(const MessageChain &mc) const;
 
         [[nodiscard]] bool empty() const;
-
-        /// @brief 回复并发送
-        /// @param s 内容
-        /// @param groupid 如果是来源于TempGroupMessage就要提供(因为要找到那个Member)
-        /// @note 可以改MessageSource里的内容, 客户端在发送的时候并不会校验MessageSource的内容正确性(比如改originalMessage来改引用的文本的内容, 或者改id来定位到其他信息)
-        /// @detail 支持以下类型传入
-        /// - std::string / const char* 相当于传入PlainText(str)
-        /// - SingleMessage的各种派生类
-        /// - MessageChain
-        /// @deprecated use Contact.quoteAndSend or `this->quoteAndSend1(s, groupId, env)`, since v2.8.1
-        template<class T>
-        ShouldNotUse("use Contact.quoteAndSend") MessageSource
-                quoteAndSendMessage(T s, QQID groupid = -1, void *env = nullptr) = delete;
 
     public: // static functions
         /// @brief 找到miraiCode结尾的`]`
@@ -245,8 +233,8 @@ namespace MiraiCP {
         }
 
         template<typename... T2>
-        void constructMessages(const std::string &h, T2 &&...args) {
-            emplace_back(PlainText(h));
+        void constructMessages(std::string h, T2 &&...args) {
+            emplace_back(PlainText(std::move(h)));
             constructMessages(std::forward<T2>(args)...);
         }
