+++ cpp/src/sdk/ForwardedMessage.cpp
@@ -16,6 +16,9 @@
 
 #include "ForwardedMessage.h"
 #include "Contact.h"
+#include "ExceptionHandle.h"
+#include "Exceptions/API.h"
+#include "Exceptions/IllegalArgument.h"
 #include "KtOperation.h"
 #include <json.hpp>
