+++ cpp/src/sdk/Friend.cpp
@@ -15,7 +15,9 @@
 //
 
 #include "Friend.h"
-#include "Exception.h"
+#include "ExceptionHandle.h"
+#include "Exceptions/IllegalArgument.h"
+#include "Exceptions/IllegalState.h"
 #include "JsonTools.h"
 #include "KtOperation.h"
 #include "LowLevelAPI.h"
