+++ cpp/src/sdk/Group.cpp
@@ -15,7 +15,11 @@
 //
 
 #include "Group.h"
-#include "Exception.h"
+#include "ExceptionHandle.h"
+#include "Exceptions/IllegalArgument.h"
+#include "Exceptions/IllegalState.h"
+#include "Exceptions/RemoteAsset.h"
+#include "Exceptions/Upload.h"
 #include "JsonTools.h"
 #include "KtOperation.h"
 #include "LowLevelAPI.h"
