+++ cpp/src/sdk/LowLevelAPI.cpp
@@ -15,7 +15,7 @@
 //
 
 #include "LowLevelAPI.h"
-#include "Exception.h"
+#include "ExceptionHandle.h"
 #include "JsonTools.h"
 #include "KtOperation.h"
 #include "Tools.h"
